testing amb sql funcionant correctament issue 23

+++ api/v1/containerService.js
@@ -6,10 +6,10 @@ const database = require(`${__base}api/database`);
 const containerService = {
 
   async getContainer(id) {
-    let sql = `SELECT * FROM container WHERE id = ${id}`;
+    let sql = `SELECT * FROM container WHERE id = ?`;
 
-    let [rows, fields] = await database.getPromise().query(sql, []);
-    console.log(rows);
+    let [rows, fields] = await database.getPromise().query(sql, [id]);
+
     return rows[0];
     
   },
@@ -23,9 +23,9 @@ const containerService = {
   },
 
   async getClientContainers(clientId) {
-    let sql = `SELECT * FROM container WHERE id = ${clientId}`;
+    let sql = `SELECT * FROM container WHERE id = ?`;
     
-    let [rows, fields] = await database.getPromise().query(sql, []);
+    let [rows, fields] = await database.getPromise().query(sql, [clientId]);
     return rows;
     //return containers.filter(c => c.clientId == clientId);
   },
@@ -33,57 +33,82 @@ const containerService = {
   async postContainer(container){
 
     let sql = `INSERT INTO container(clientId, code, description, width, length, height, maxWeight)
-                VALUES('${container.clientId}', '${container.code}', '${container.description}', '${container.width}',
-                 '${ container.length }', '${container.height}', '${container.maxWeight}')`;
+              VALUES(?, ?, ?, ?, ?, ?, ?)`;
+    
+    let values = [
+      container.clientId,
+      container.code,
+      container.description,
+      container.width,
+      container.length,
+      container.height,
+      container.maxWeight
+    ];
 
-    let [rows, fields] = await database.getPromise().query(sql, []);
-    return rows;
-    // const nextId = containers.reduce((maxId, container) => Math.max(maxId, container.id), 0) + 1;
-    // containers.push({ ...container, id: nextId });
-    // return containers[containers.length-1];
+    let [rows, fields] = await database.getPromise().query(sql, values);
+
+    sql = `SELECT * FROM container WHERE id = ?`;
+    [rows, fields] = await database.getPromise().query(sql, [rows.insertId]);
+
+    return rows[0];
 
   },
 
   async putContainer(id, newContainerData) {
 
-    let sql = `UPDATE container SET
-      clientId = '${newContainerData.clientId}',
+    sql = `UPDATE container SET
       code = '${newContainerData.code}',
       description = '${newContainerData.description}',
       width = '${newContainerData.width}',
-      length = '${ newContainerData.length }',
+      length = '${newContainerData.length}',
       height = '${newContainerData.height}',
       maxWeight = '${newContainerData.maxWeight}'
-    WHERE id = ${ id }`;
+    WHERE id = ${id}`;
 
     let [rows, fields] = await database.getPromise().query(sql, []);
-    return rows;
-    // const containerToUpdate = containers.find(container => container.id == id);
-    // if (containerToUpdate) {
-    //   containerToUpdate.clientId = newContainerData.clientId || containerToUpdate.clientId;
-    //   containerToUpdate.code = newContainerData.code || containerToUpdate.code;
-    //   containerToUpdate.description = newContainerData.description || containerToUpdate.description;
-    //   containerToUpdate.width = newContainerData.width || containerToUpdate.width;
-    //   containerToUpdate.length = newContainerData.length || containerToUpdate.length;
-    //   containerToUpdate.height = newContainerData.height || containerToUpdate.height;
-    //   containerToUpdate.maxWeight = newContainerData.maxWeight || containerToUpdate.maxWeight;
-    // }
-    // return containerToUpdate;
+
+    // return undefined if client not found
+    if (rows.affectedRows === 0) {
+      return undefined;
+    }
+
+    if (rows.affectedRows !== 1) {
+      throw new Error(`Error updating client, affected rows = ${rows.affectedRows}`);
+    }
+
+    sql = `SELECT * FROM container WHERE id = ${id}`;
+    [rows, fields] = await database.getPromise().query(sql, []);
+
+    if (rows.length !== 1) {
+      throw new Error(`Error retrieving updated client data`);
+    }
+
+    return rows[0];
+
   },
 
   async deleteContainer(id) {
 
-    let sql = `DELETE FROM container WHERE id = ${ id }`;
+    let sql = `SELECT * FROM container WHERE id = ?`;
+    let [rows, fields] = await database.getPromise().query(sql, [id]);
 
-    let [rows, fields] = await database.getPromise().query(sql, []);
-    return rows;
-    // const index = containers.findIndex(o => o.id == id); 
-    // if (index >= 0) {
-    //   let conatinerDeleted = containers.splice(index, 1); 
-    //   return conatinerDeleted[0];
-    // } else {
-    //   return undefined; 
-    // }
+    if (rows.length !== 1) {
+      
+      return undefined;
+    }
+
+    let containerToDelete = rows[0];
+    
+    sql = `DELETE FROM container WHERE id = ${ id }`;
+
+    [rows, fields] = await database.getPromise().query(sql, []);
+
+    if (rows.affectedRows !== 1) {
+      throw new Error(`Error deleting container, affected rows = ${rows.affectedRows}`);
+    }
+     
+    return containerToDelete;
+
   }
 
 };
